+++ apps/docs/app/examples/cards/_components/payment-method.tsx
@@ -7,7 +7,7 @@ import { type JSX } from 'react';
 import { TextInput } from '@codefast/ui/text-input';
 import { Icons } from '@/components/icons';
 
-export function DemoPaymentMethod(): JSX.Element {
+export function PaymentMethod(): JSX.Element {
   return (
     <Card>
       <CardHeader>