+++ apps/docs/app/examples/cards/page.tsx
@@ -2,22 +2,22 @@ import { cn } from '@codefast/ui/utils';
 import Image from 'next/image';
 import { type HTMLAttributes, type JSX } from 'react';
 import { type Metadata } from 'next';
-import { DemoCookieSettings } from '@/app/examples/cards/_components/cookie-settings';
-import { DemoCreateAccount } from '@/app/examples/cards/_components/create-account';
-import { DemoDatePicker } from '@/app/examples/cards/_components/date-picker';
-import { DemoGithub } from '@/app/examples/cards/_components/github-card';
-import { DemoNotifications } from '@/app/examples/cards/_components/notifications';
-import { DemoPaymentMethod } from '@/app/examples/cards/_components/payment-method';
-import { DemoReportAnIssue } from '@/app/examples/cards/_components/report-an-issue';
-import { DemoShareDocument } from '@/app/examples/cards/_components/share-document';
-import { DemoTeamMembers } from '@/app/examples/cards/_components/team-members';
+import { CookieSettings } from '@/app/examples/cards/_components/cookie-settings';
+import { CreateAccount } from '@/app/examples/cards/_components/create-account';
+import { DatePicker } from '@/app/examples/cards/_components/date-picker';
+import { Github } from '@/app/examples/cards/_components/github-card';
+import { Notifications } from '@/app/examples/cards/_components/notifications';
+import { PaymentMethod } from '@/app/examples/cards/_components/payment-method';
+import { ReportAnIssue } from '@/app/examples/cards/_components/report-an-issue';
+import { ShareDocument } from '@/app/examples/cards/_components/share-document';
+import { TeamMembers } from '@/app/examples/cards/_components/team-members';
 
 export const metadata: Metadata = {
   title: 'Cards',
   description: 'Examples of cards built using the components.',
 };
 
-function DemoContainer({ className, ...props }: HTMLAttributes<HTMLDivElement>): JSX.Element {
+function Container({ className, ...props }: HTMLAttributes<HTMLDivElement>): JSX.Element {
   return <div className={cn('flex items-center justify-center [&>div]:w-full', className)} {...props} />;
 }
 
@@ -30,37 +30,37 @@ export default function CardsPage(): JSX.Element {
       </div>
       <div className="hidden items-start justify-center gap-6 rounded-lg p-8 md:grid lg:grid-cols-2 xl:grid-cols-3">
         <div className="col-span-2 grid items-start gap-6 lg:col-span-1">
-          <DemoContainer>
-            <DemoCreateAccount />
-          </DemoContainer>
-          <DemoContainer>
-            <DemoPaymentMethod />
-          </DemoContainer>
+          <Container>
+            <CreateAccount />
+          </Container>
+          <Container>
+            <PaymentMethod />
+          </Container>
         </div>
         <div className="col-span-2 grid items-start gap-6 lg:col-span-1">
-          <DemoContainer>
-            <DemoTeamMembers />
-          </DemoContainer>
-          <DemoContainer>
-            <DemoShareDocument />
-          </DemoContainer>
-          <DemoContainer>
-            <DemoDatePicker />
-          </DemoContainer>
-          <DemoContainer>
-            <DemoNotifications />
-          </DemoContainer>
+          <Container>
+            <TeamMembers />
+          </Container>
+          <Container>
+            <ShareDocument />
+          </Container>
+          <Container>
+            <DatePicker />
+          </Container>
+          <Container>
+            <Notifications />
+          </Container>
         </div>
         <div className="col-span-2 grid items-start gap-6 lg:col-span-2 lg:grid-cols-2 xl:col-span-1 xl:grid-cols-1">
-          <DemoContainer>
-            <DemoReportAnIssue />
-          </DemoContainer>
-          <DemoContainer>
-            <DemoGithub />
-          </DemoContainer>
-          <DemoContainer>
-            <DemoCookieSettings />
-          </DemoContainer>
+          <Container>
+            <ReportAnIssue />
+          </Container>
+          <Container>
+            <Github />
+          </Container>
+          <Container>
+            <CookieSettings />
+          </Container>
         </div>
       </div>
     </>