+++ packages/ui/src/components/checkbox-cards.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import * as React from 'react';
 import { CheckIcon } from '@radix-ui/react-icons';
 import * as CheckboxGroupPrimitive from '@codefast-ui/checkbox-group';