+++ packages/ui/src/components/menubar.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import * as React from 'react';
 import { CheckIcon, ChevronRightIcon, DotFilledIcon } from '@radix-ui/react-icons';
 import * as MenubarPrimitive from '@radix-ui/react-menubar';