+++ packages/ui/src/components/text-input.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import * as React from 'react';
 import * as InputPrimitive from '@codefast-ui/input';
 import { Spinner } from '@/components/spinner';